Allow snis_update_assets.sh to copy files locally

+++ util/snis_update_assets.sh
@@ -9,7 +9,19 @@ new_count=0
 update_count=0
 update_fail_count=0
 DESTDIR=.
+FETCH="fetch"
+localcopy=0
 export dryrun=0
+
+usage()
+{
+	echo "Usage:"
+	echo "snis_update_assets.sh"
+	echo "snis_update_assets.sh --localcopy --destdir some-directory"
+	echo "snis_update_assets.sh --dry-run"
+	echo "snis_update_assets.sh --dry-run --localcopy --destdir some-directory"
+	exit 0
+}
 
 sanity_check_environment()
 {
@@ -26,7 +38,9 @@ fetch_file()
 	local URL="$1"
 	local FILE="$2"
 	local dryrun="$3"
-	updating_or_creating="$4"
+	local updating_or_creating="$4"
+	local localcopy="$5"
+
 	if [ "$dryrun" != "0" ]
 	then
 		updating_or_creating="Not $updating_or_creating"
@@ -34,13 +48,30 @@ fetch_file()
 	echo -n 1>&2 "$updating_or_creating $FILE... "
 	if [ "$dryrun" = "0" ]
 	then
-		wget --quiet "$1" -O - > "$FILE"
-		if [ "$?" != "0" ]
+		if [ "$localcopy" = 0 ]
 		then
-			/bin/rm -f "$FILE"
-			echo 1>&2
-			echo "$PROG"': Failed to fetch '"$URL" 1>&2
-			return 1
+			wget --quiet "$1" -O - > "$FILE"
+			if [ "$?" != "0" ]
+			then
+				/bin/rm -f "$FILE"
+				echo 1>&2
+				echo "$PROG"': Failed to fetch '"$URL" 1>&2
+				return 1
+			fi
+		else
+			if [ "$1" = "$FILE" ]
+			then
+				echo "$1 and $FILE are the same." 1>&2
+			else
+				/bin/cp "$1" "$FILE"
+				if [ "$?" != "0" ]
+				then
+					/bin/rm -f "$FILE"
+					echo 1>&2
+					echo "$PROG"': Failed to copy '"$URL" 1>&2
+					return 1
+				fi
+			fi
 		fi
 	fi
 	echo "done" 1>&2
@@ -63,6 +94,7 @@ update_file()
 {
 	local checksum="$1"
 	local filename="$2"
+	local localcopy="$3"
 	if [ -f "$DESTDIR/$filename" ]
 	then
 		localchksum=$(md5sum "$DESTDIR"/"$filename" | awk '{ print $1 }')
@@ -76,7 +108,7 @@ update_file()
 				echo "$PROG"':Cannot move old '"$DESTDIR"/"$filename"' out of the way, skipping' 1>&2
 				update_fail_count=$((update_fail_count + 1))
 			else
-				fetch_file "$ASSET_URL"/"$filename" "$DESTDIR"/"$filename" "$dryrun" "Updating"
+				fetch_file "$ASSET_URL"/"$filename" "$DESTDIR"/"$filename" "$dryrun" "Updating" $localcopy 
 				if [ "$?" != "0" ]
 				then
 					update_fail_count=$((update_fail_count + 1))
@@ -101,7 +133,7 @@ update_file()
 		fi
 		if [ -d "$dname" -o "$dryrun" != "0" ]
 		then
-			fetch_file $ASSET_URL/$filename "$DESTDIR"/"$filename" "$dryrun" "Creating"
+			fetch_file $ASSET_URL/$filename "$DESTDIR"/"$filename" "$dryrun" "Creating" $localcopy
 			if [ "$?" = "0" ]
 			then
 				new_count=$((new_count+1))
@@ -122,7 +154,7 @@ update_files()
 		then
 			break;
 		fi
-		update_file $x
+		update_file $x $localcopy
 	done < "$MANIFEST"
 }
 
@@ -145,10 +177,25 @@ output_statistics()
 	echo
 }
 
+if [ "$1" = "--help" ]
+then
+	usage;
+fi
+
 if [ "$1" = "--dry-run" ]
 then
 	dryrun=1
 	shift;
+fi
+
+if [ "$1" = "--localcopy" ]
+then
+	shift;
+	ASSET_URL="./"
+	localcopy=1
+	MANIFEST_URL=./share/snis/original_manifest.txt
+	FETCH="copy"
+	echo "Local copy mode enabled"
 fi
 
 if [ "$1" = "--destdir" ]
@@ -158,8 +205,14 @@ then
 	echo "Using DESTDIR=$DESTDIR"
 fi
 
+if [ "$localcopy" = "1" -a "$DESTDIR" = "." ]
+then
+	echo "Cannot localcopy from . to ." 1>&2
+	exit 1
+fi
+
 sanity_check_environment
-fetch_file "$MANIFEST_URL" "$MANIFEST_FILE" 0 "Fetching"
+fetch_file "$MANIFEST_URL" "$MANIFEST_FILE" 0 "$FETCH""ing" $localcopy
 if [ "$?" != "0" ]
 then
 	exit 1
@@ -167,5 +220,6 @@ fi
 update_files "$MANIFEST_FILE"
 output_statistics
 
+/bin/cp "$MANIFEST_FILE" "$DESTDIR"/share/snis/original_manifest.txt
 /bin/rm "$MANIFEST_FILE"
